requirements.txt
flask>=2.3.0,<3.0.0
werkzeug>=2.3.0,<3.0.0
gunicorn>=21.0.0
weasyprint>=60.0,<63.0
matplotlib>=3.7.0,<4.0.0
anthropic>=0.25.0
Pillow>=10.0.0
python-dotenv>=1.0.0
